Type 3 functionality added

+++ src/main/java/comp307/backend/booking/Service/BookingService.java
@@ -8,6 +8,7 @@ import comp307.backend.account.Object.User;
 import comp307.backend.account.Object.UserRepository;
 import comp307.backend.booking.Entity.Booking;
 import comp307.backend.booking.Entity.BookingSlot;
+import comp307.backend.booking.Entity.GroupMeetingInstance;
 import comp307.backend.booking.Repository.*;
 
 import java.time.LocalDateTime;
@@ -18,11 +19,13 @@ import java.util.List;
 public class BookingService {
     private final BookingRepository bookingRepository;
     private final BookingSlotRepository bookingSlotRepository;
+    private final GroupMeetingInstanceRepository groupMeetingInstanceRepository;
     private final UserRepository userRepository;
 
-    public BookingService(BookingRepository bookingRepository, BookingSlotRepository bookingSlotRepository, UserRepository userRepository) {
+    public BookingService(BookingRepository bookingRepository, BookingSlotRepository bookingSlotRepository, GroupMeetingInstanceRepository groupMeetingInstanceRepository, UserRepository userRepository) {
         this.bookingRepository = bookingRepository;
         this.bookingSlotRepository = bookingSlotRepository;
+        this.groupMeetingInstanceRepository = groupMeetingInstanceRepository;
         this.userRepository = userRepository;
     }
 
@@ -51,13 +54,16 @@ public class BookingService {
 
     //Type 2
     //Already enforced on front end that there can't be overlapping groups or end date earlier than start date etc.
-    public BookingSlot createGroupBookingSlot() {
-        //todo
-        throw new RuntimeException("Not implemented yet");
+    public BookingSlot createGroupMeetingBookingProposalSlot(Long groupMeetingInstanceID, LocalDateTime startDateTime, LocalDateTime endDateTime) {
+        GroupMeetingInstance groupMeetingInstance = groupMeetingInstanceRepository.findById(groupMeetingInstanceID).orElseThrow(() -> new RuntimeException("Group meeting instance " + groupMeetingInstanceID + " not found."));
+
+        BookingSlot bookingSlot = new BookingSlot(groupMeetingInstance.getOwner(), startDateTime, endDateTime, groupMeetingInstance);
+        return bookingSlotRepository.save(bookingSlot);
+        
     }
 
     public List<BookingSlot> getAllOwnedSlots(String ownerEmail) {
-        User owner = userRepository.findById(ownerEmail).orElseThrow(() -> new RuntimeException("User " + ownerEmail + " not found."));
+        User owner = userRepository.findById(ownerEmail).orElseThrow(() -> new IllegalArgumentException("User " + ownerEmail + " not found."));
 
         return bookingSlotRepository.findByOwner(owner);
         
@@ -69,7 +75,7 @@ public class BookingService {
 
     //maybe add email service, or frontend could default open email with all the people whose bookings got cancelled in which case can return list of emails that should be notified
     public void cancelBookingSlot(Long bookingSlotId) {
-        BookingSlot bookingSlot = bookingSlotRepository.findById(bookingSlotId).orElseThrow(() -> new RuntimeException("Slot " + bookingSlotId + " not found."));
+        BookingSlot bookingSlot = bookingSlotRepository.findById(bookingSlotId).orElseThrow(() -> new IllegalArgumentException("Slot " + bookingSlotId + " not found."));
         bookingSlot.setSlotStatus(BookingSlot.BookingSlotStatus.CANCELLED);
         bookingSlotRepository.save(bookingSlot);
 
@@ -83,43 +89,58 @@ public class BookingService {
 
 
     //BOOKING
+    //For type 3
     public Booking book(Long bookingSlotId, String reserveeEmail) {
         User reservee = userRepository.findById(reserveeEmail).orElseThrow(() -> new RuntimeException("User " + reserveeEmail + " not found."));
         BookingSlot bookingSlot = bookingSlotRepository.findById(bookingSlotId).orElseThrow(() -> new RuntimeException("Slot " + bookingSlotId + " not found."));
 
         if (!bookingSlot.getSlotStatus().equals(BookingSlot.BookingSlotStatus.AVAILABLE)) {
-            throw new RuntimeException("Slot " + bookingSlotId + " is not available.");
+            throw new IllegalArgumentException("Slot " + bookingSlotId + " is not available.");
         }
 
-        if (bookingSlot.getType().equals(BookingSlot.BookingSlotType.OFFICE_HOURS)) {
-            //can never be full so still available.
-            return bookingRepository.save(new Booking(bookingSlot, reservee));
+        if (bookingSlot.getSlotType() != BookingSlot.BookingSlotType.OFFICE_HOURS) {
+            throw new IllegalArgumentException("Slot " + bookingSlotId + " is not an office hours slot.");
         }
-        //else it's a group slot and it's limited
-        else {
-            int currentBookings = bookingRepository.findByBookingSlot(bookingSlot).size();
 
-            if (currentBookings == bookingSlot.getMaxUsers()) {
-                //maybe don't throw
-                throw new RuntimeException("Slot " + bookingSlotId + " is full.");
-            }
-            else {
-                if (currentBookings + 1 == bookingSlot.getMaxUsers()) {
-                    bookingSlot.setSlotStatus(BookingSlot.BookingSlotStatus.BOOKED);
-                    bookingSlotRepository.save(bookingSlot);
-                }
+        //can never be full so still available.
+        return bookingRepository.save(new Booking(bookingSlot, reservee));
+        
+    }
 
-                return bookingRepository.save(new Booking(bookingSlot, reservee));
-            }
+    //Type 2
+    //Might need another function/modify this one to add the ability to join the Selected group meeting proposal that they haven't accepted yet.
+    public Booking markAvailabilityForProposal(Long bookingSlotId, String reserveeEmail) {
+        User reservee = userRepository.findById(reserveeEmail).orElseThrow(() -> new RuntimeException("User " + reserveeEmail + " not found."));
+        BookingSlot bookingSlot = bookingSlotRepository.findById(bookingSlotId).orElseThrow(() -> new RuntimeException("Slot " + bookingSlotId + " not found."));
+
+        if (bookingSlot.getSlotType() != BookingSlot.BookingSlotType.GROUP_PROPOSAL) {
+            throw new IllegalArgumentException("Slot " + bookingSlotId + " is not a pending proposal slot.");
         }
+
+        if (bookingSlot.getSlotStatus() != BookingSlot.BookingSlotStatus.AVAILABLE) {
+            throw new IllegalArgumentException("Slot " + bookingSlotId + " is not available for marking availability.");
+        }
+
+        int currentBookingsCount = bookingRepository.findByBookingSlot(bookingSlot).size();
+        if (currentBookingsCount + 1 == bookingSlot.getMaxUsers()) {
+            bookingSlot.setSlotStatus(BookingSlot.BookingSlotStatus.BOOKED);
+            bookingSlotRepository.save(bookingSlot);
+        }
+
+        return bookingRepository.save(new Booking(bookingSlot, reservee));
     }
 
     //whether its type 2 or type 3, either way the booking will become available when unbooked because it either had infinite space or now has at least 1 space
     public void unbook(Long bookingId) {
         Booking booking = bookingRepository.findById(bookingId).orElseThrow(() -> new RuntimeException("Booking " + bookingId + " not found."));
         BookingSlot bookingSlot = booking.getBookingSlot();
+
+        if (bookingSlot.getSlotStatus() == BookingSlot.BookingSlotStatus.CANCELLED) {
+            throw new IllegalArgumentException("Slot " + bookingSlot.getBookingSlotID() + " is cancelled, should not be calling this function.");
+        }
+
         bookingSlot.setSlotStatus(BookingSlot.BookingSlotStatus.AVAILABLE);
         bookingSlotRepository.save(bookingSlot);
         bookingRepository.delete(booking);
-    }   
+    }
 }
